+++ packages/megadoc-html-serializer/lib/index.js
@@ -1,1 +1,3 @@
 module.exports = require('./HTMLSerializer');
+module.exports.MediaWikiLinkStrategy = require('./render/LinkResolver__MediaWikiLinkStrategy')
+module.exports.MegadocLinkStrategy = require('./render/LinkResolver__MegadocLinkStrategy')